+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+- Add levelling system household check toggle for compatibility with Realistic Workplaces and Households mod.
+
 ### 0.8.6
 - Update UI dependency versions to match game version 1.5.2.
 - Re-enable burst compilation of building levelling code for game version 1.5.2.